requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
jinja2
passlib[bcrypt]
python-jose[cryptography]
python-multipart
requests
